move ProgressBar for register into register page
