+++ src/app/(post-login)/submission/program/[shortName]/registration/components/ProgressBar.tsx
